+++ console/migrations/m210405_005233_add_tax_percents_to_dv_aucs_table.php
@@ -12,12 +12,12 @@ class m210405_005233_add_tax_percents_to_dv_aucs_table extends Migration
      */
     public function safeUp()
     {
-        $this->addColumn('dv_aucs', 'one_percent_ewt', $this->float());
-        $this->addColumn('dv_aucs', 'two_percent_ewt', $this->float());
-        $this->addColumn('dv_aucs', 'three_percent_ft', $this->float());
-        $this->addColumn('dv_aucs', 'five_percent_ft', $this->float());
-        $this->addColumn('dv_aucs', 'five_percent_ewt', $this->float());
-        $this->addColumn('dv_aucs', 'total_withheld', $this->float());
+        $this->addColumn('dv_aucs', 'one_percent_ewt', $this->decimal(10,2));
+        $this->addColumn('dv_aucs', 'two_percent_ewt', $this->decimal(10,2));
+        $this->addColumn('dv_aucs', 'three_percent_ft', $this->decimal(10,2));
+        $this->addColumn('dv_aucs', 'five_percent_ft', $this->decimal(10,2));
+        $this->addColumn('dv_aucs', 'five_percent_ewt', $this->decimal(10,2));
+        $this->addColumn('dv_aucs', 'total_withheld', $this->decimal(10,2));
     }
 
     /**
